\name{R2STATS}
\alias{R2STATS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
R2STATS
}
\description{
A function to construct and launch the R2STATS interface.
}
\usage{
R2STATS()
}
\details{
The interface is stored in the proto object 'r2stats'.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
  \item{r2stats}{A proto object which field 'models' contains the list of all models tested, in R2STATS model representation.}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\author{
Yvonnick Noel, University of Brittany, Rennes, France.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ GUI }

