\name{R2STATS-package}
\alias{R2STATS-package}
\docType{package}
\title{
A gWidgetsRGtk2 Graphical User Interface for GLM and GLMM in R
}
\description{
R2STATS is a graphical frontend to R designed to make estimation and visualization of GLM and GLMM easy.
}
\details{
\tabular{ll}{
Package: \tab R2STATS\cr
Type: \tab Package\cr
Version: \tab 0.68-28\cr
Date: \tab 2011-11-07\cr
License: \tab GPL (>=2)\cr
}
Just type R2STATS() at the prompt to launch the GUI.
}
\author{
Maintainer: Yvonnick Noel <yvonnick.noel@uhb.fr>
}

